pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fruit-cnn-ml-project"
version = "0.1.0"
description = "CNN training and evaluation pipeline for fruit image classification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "tensorflow>=2.13",
  "matplotlib>=3.7",
  "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[tool.setuptools]
packages = ["src", "src.data", "src.features", "src.models", "src.utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
